doc/integrations/cortx-s3-slack-bot/requirements.txt
aiohttp==3.7.4.post0
argcomplete==1.10.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.8.0
boto3==1.17.57
botocore==1.20.57
certifi==2020.12.5
chardet==3.0.4
docx2txt==0.8
EbookLib==0.17.1
elasticsearch==7.12.0
extract-msg==0.23.1
idna==2.10
IMAPClient==2.1.0
jmespath==0.10.0
lxml==4.6.3
multidict==5.1.0
olefile==0.46
pdfminer.six==20181108
Pillow==8.2.0
pycryptodome==3.10.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-pptx==0.6.18
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.12.0
slack-bolt==1.5.0
slack-sdk==3.5.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
textract==1.6.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.4
xlrd==1.2.0
XlsxWriter==1.4.0
yarl==1.6.3
